+++ src/viewmodels/DovetCoversViewModel.java
@@ -10,8 +10,8 @@ public class DovetCoversViewModel extends AbstractViewModel{
     private Double wastage = 0.0;
     private Double flap = 0.0;
     private boolean isCustom = false;
-    private int customWidth = 0;
-    private int customHeight = 0;
+    private Double customWidth = 0.0;
+    private Double customHeight = 0.0;
     private boolean includeLable = false;
     private boolean includeTag = false;
     private boolean includeSealBag = false;
@@ -122,19 +122,19 @@ public class DovetCoversViewModel extends AbstractViewModel{
         this.cardboardCost = cardboardCost;
     }
 
-    public int getCustomHeight() {
+    public Double getCustomHeight() {
         return customHeight;
     }
 
-    public void setCustomHeight(int customHeight) {
+    public void setCustomHeight(Double customHeight) {
         this.customHeight = customHeight;
     }
 
-    public int getCustomWidth() {
+    public Double getCustomWidth() {
         return customWidth;
     }
 
-    public void setCustomWidth(int customWidth) {
+    public void setCustomWidth(Double customWidth) {
         this.customWidth = customWidth;
     }
