+++ src/viewmodels/PillowcaseViewModel.java
@@ -10,8 +10,8 @@ public class PillowcaseViewModel extends AbstractViewModel{
     private boolean isCustom = false;
     private Double flap = 0.0;
     private Double frill = 0.0;
-    private int customWidth = 0;
-    private int customHeight = 0;
+    private Double customWidth = 0.0;
+    private Double customHeight = 0.0;
     private boolean includeLable = false;
     private boolean includeTag = false;
     private boolean includeSealBag =  false;
@@ -73,19 +73,19 @@ public class PillowcaseViewModel extends AbstractViewModel{
         this.cardboardCost = cardboardCost;
     }
 
-    public int getCustomHeight() {
+    public Double getCustomHeight() {
         return customHeight;
     }
 
-    public void setCustomHeight(int customHeight) {
+    public void setCustomHeight(Double customHeight) {
         this.customHeight = customHeight;
     }
 
-    public int getCustomWidth() {
+    public Double getCustomWidth() {
         return customWidth;
     }
 
-    public void setCustomWidth(int customWidth) {
+    public void setCustomWidth(Double customWidth) {
         this.customWidth = customWidth;
     }
